+++ R/att_from_rmds.R
@@ -94,7 +94,7 @@ att_from_rmd <- function(path, temp_dir = tempdir(), warn = -1,
 att_from_rmds <- function(path = "vignettes",
                           pattern = "*.[.](Rmd|rmd|qmd)$",
                           recursive = TRUE, warn = -1,
-                          inside_rmd = FALSE, inline = TRUE) {
+                          inside_rmd = FALSE, inline = TRUE,folder_to_exclude = "renv") {
 
   if (isTRUE(all(dir.exists(path)))) {
     all_f <- list.files(path, full.names = TRUE, pattern = pattern, recursive = recursive)
@@ -103,6 +103,17 @@ att_from_rmds <- function(path = "vignettes",
   } else {
     stop("Some files/directories do not exist")
   }
+
+  if (!is.null(folder_to_exclude) && length(folder_to_exclude) > 0) {
+    exclude_files <- unlist(lapply(folder_to_exclude, function(folder) {
+      list.files(path = file.path(path, folder), full.names = TRUE, pattern = pattern, recursive = recursive)
+    }))
+
+    # Exclure les fichiers
+    all_f <- setdiff(all_f, exclude_files)
+  }
+
+
 
   res <- lapply(all_f,
                 function(x) att_from_rmd(