+++ tests/testthat/f3.R
@@ -1,0 +1,1 @@
+library(dont.find.me3)